Safari-609.1.11.1/LayoutTests/fast/events/click-focus-control-expected.txt

This test ensures that we can click to focus an a element. Click on the element below.

The expected result is platform specific. Mac doesn't allow some form controls to be mouse focusable.

Result

a1 received focus ()
a2 received focus ()
a3 received focus ()
a4 received focus ()
a5 received focus ()
PASS All form elements got focus
